Added Intro&Menu & Big fix to road
Intro and Menu added, not yet fully functional though.

Big fix on how a truck can reverse and denied the player the ability to
rotate a road when a truck is over it. Trucks should no longer continue
straight when next road get rotated and the truck is very close to it,
instead the truck reverse, since it isnt allowed.