Victory Major Update 1
Attention: Work in progress
- Added Level conditions.
- Added Level condition array to Profile Manager (DontDestroyOnLoad
Singleton)
- Added Several Classes for LevelCondition to Work, Specially class
"Quest"
- Quest: misions to do, returns stars. (usually 3 of this per level, 1
star, 2 stars, 3 stars).
- Added GameModes. Currently 1 game mode: TimeAttack
- Added level management methods on GameController. Specially Events
that triggers a listener depending on the GameMode. This events Updates
Quests when "Score" happens, or when player gains money.

TimeAttack mode: Level management methods register an Action on timer's
end at the begining of the level. When timer is finish that Action Runs
and check how many stars were achieved. For prompts it to Debug.log

NO GUI YET